Add locking on block basis for user-facing functions and prettify contracts

+++ contracts/test/DAI.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.7.6;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract DAI is ERC20 {
-    constructor (string memory name, string memory symbol) ERC20(name, symbol) {
-        _mint(msg.sender, 1000000000000000 * 10 ** uint(decimals()));
+    constructor(string memory name, string memory symbol) ERC20(name, symbol) {
+        _mint(msg.sender, 1000000000000000 * 10**uint256(decimals()));
     }
 }
